+++ projects/igniteui-angular/src/lib/grids/grid/grid-filtering.spec.ts
@@ -605,11 +605,18 @@ describe('IgxGrid - Filtering actions', () => {
         }).toThrowError('Invalid condition or Expression Tree!');
     }));
 
-    it('Should throw descriptive error when filterGlobal() is called without condition', fakeAsync(() => {
-        expect(() => {
-            grid.filterGlobal(100);
-            fix.detectChanges();
-        }).toThrowError('Invalid condition!');
+    it('Should not clear previous filtering when filterGlobal() is called with invalid condition', fakeAsync(() => {
+        grid.filter('Downloads', 100, IgxNumberFilteringOperand.instance().condition('greaterThan'), true);
+        fix.detectChanges();
+        expect(grid.rowList.length).toEqual(4);
+        expect(grid.getCellByColumn(0, 'Downloads').value).toEqual(254);
+
+        // Execute global filtering with invalid condition.
+        grid.filterGlobal(1000, null);
+        fix.detectChanges();
+
+        expect(grid.rowList.length).toEqual(4);
+        expect(grid.getCellByColumn(0, 'Downloads').value).toEqual(254);
     }));
 });