+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes for each version of this project will be documented in this 
 
 ## 7.2.11
 - `IgxGrid`, `IgxTreeGrid`, `IgxHierarchicalGrid`
-    - **Breaking Change** The **condition** parameter of the `filterGlobal` method is no longer optional. When the filterGlobal method is called with an invalid condition, it will now throw an error and will not clear the existing filters for all columns.
+    - **Breaking Change** The **condition** parameter of the `filterGlobal` method is no longer optional. When the filterGlobal method is called with an invalid condition, it will not clear the existing filters for all columns.
 
 ## 7.2.9
 - `Pager`